fix: stability gates — remove unnecessary cast, exclude test infra from unwrap scan
- krun/process.rs: remove redundant `sig as i32` cast (clippy)
- stability-gates.yml: exclude minibox-core/src/adapters/ (test-only infra
  behind test-utils feature) and minibox-testers from unwrap scan

+++ crates/macbox/src/krun/process.rs
@@ -63,7 +63,7 @@ impl SmolvmProcess {
             None => {
                 let sig = status.signal().unwrap_or(0);
                 tracing::debug!(signal = sig, "smolvm: process killed by signal");
-                Ok(-(sig as i32))
+                Ok(-sig)
             }
         }
     }
